fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9463 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBNode.cpp
@@ -147,7 +147,7 @@ NBNode::ApproachingDivider::spread(const std::vector<int> &approachingLanes,
         //
         // check whether the left boundary of the approached street has
         //  been overridden; if so, move all lanes to the right
-        if (dest+loffset >= noOutgoingLanes) {
+        if (dest+loffset >= static_cast<int>(noOutgoingLanes)) {
             loffset -= 1;
             roffset += 1;
             for (unsigned int i=0; i<ret->size(); i++) {
